Refactor to move cleaning functions to utilities file
